Prep commit

+++ models/taskModel.js
@@ -25,11 +25,19 @@ export const addTaskDB = async (title, description, priority) => {
 
 export const editTaskDB = async (id, title, description, priority) => {
     try{
-        const result = await query(
-            "UPDATE tasks SET title = $1, description = $2, priority = $3 WHERE id = $4 RETURNING *",
-            [title, description, priority, id]
+        const task = await query(
+            "SELECT * FROM tasks WHERE id = $1 LIMIT 1", [id]
         );
-        return result.rows[0];
+
+        if(task.rows.length !== 0){
+            const result = await query(
+                "UPDATE tasks SET title = $1, description = $2, priority = $3 WHERE id = $4 RETURNING *",
+                [title, description, priority, id]
+            );
+            return result.rows[0];
+        }else{
+            throw new Error("Task not found.");
+        }
     }catch(error){
         console.error("Error editing task:", error);
         throw error;
@@ -38,10 +46,18 @@ export const editTaskDB = async (id, title, description, priority) => {
 
 export const toggleTaskDB = async (task_id) => {
     try{
-        const result = await query(
-            "UPDATE tasks SET completed = NOT completed WHERE id = $1 RETURNING *", [task_id]
+        const task = await query(
+            "SELECT * FROM tasks WHERE id = $1 LIMIT 1", [task_id]
         );
-        return result.rows[0];
+
+        if(task.rows.length !== 0){
+            const result = await query(
+                "UPDATE tasks SET completed = NOT completed WHERE id = $1 RETURNING *", [task_id]
+            );
+            return result.rows[0];
+        }else{
+            throw new Error("Task not found.");
+        }
     }catch(error){
         console.error("Error updating task:", error);
         throw error;
@@ -50,10 +66,18 @@ export const toggleTaskDB = async (task_id) => {
 
 export const deleteTaskDB = async (task_id) => {
     try{
-        const result = await query(
-            "DELETE FROM tasks WHERE id = $1 RETURNING *", [task_id]
+        const task = await query(
+            "SELECT * FROM tasks WHERE id = $1 LIMIT 1", [task_id]
         );
-        return result.rows[0];
+
+        if(task.rows.length !== 0){
+            const result = await query(
+                "DELETE FROM tasks WHERE id = $1 RETURNING *", [task_id]
+            );
+            return result.rows[0];
+        }else{
+            throw new Error("Task not found.");
+        }
     }catch(error){
         console.error("Error deleting task:", error);
         throw error;
